Operel/new configparser (#91)
The lion part of the reformat finally breaks backwards compatibility to
introduce the new config parser and clear separation to apps.

1. Be aware of what got deprecated:
* `main.py` and `main_interactive.py`
* Everything under `configs.yaml` is no longer relevant, apps have their
own configs
* Similarly, configs that existed under `examples/funny_nerf` were
deprecated
* The templates folder 
2. `config_parser.py` heavily revised.
3. `app/nerf`, `app/nglod` - each app now has it's own argument parsing,
additional `requirements.txt` (if needed), README, and so forth. Args
aim to be as close as possible to what we had before this change. Some
args were removed, and defaults may have been re-set to different values
(usually, if this is done it's due to recent commits which updated the
configs, or due to a parameter deemed unused).
4. `examples/` folder - both the SIGG demo and SPC browser are now more
concise, as they should be :)
5. additional changes: tracers no longer taking kwargs (on purpose),
`cuda_guard` fixed by Clement to avoid interference with the gui
libraries own argparse, the SDFDataset having a different signature

Authors: Clement Fuji-Tsang, Or Perel
Signed-off-by: operel <operel@nvidia.com>