Fixing unexpected processing of iterate() of BaseTrainer (#163)
Hi, thanks for developing such an amazing library. This is a PR to fix
the unexpected behaviour of `iterate()`. Please refer to the following
description. I hope this request will be approved.

## Changes proposed in this PR
* Initial value for `self.iteration` of BaseTrainer in the `stable`
branch.
* Processing flow inside `iterate()` of the class

## Reason for the proposal
The original iterate() calls `step()` before `pre_training()`, and
`post_training()` before and after the last `step()`.

## Reference
BaseTrainer in the `main` branch